Hopefully the tests also pass in the CI...
Split up the tests

test_s3server.py doesn't work in CI yet and s3_mock_keys_set is a
dirty workarround to pass the CI tests. Ideally I would like to use
${{ secrets.AWS_ACCESS_KEY_ID }} everywhere rather than using the
defaults for minio.

minio things only in linux

updated environment variables

access key was not available...

Use the same environment variables everywhere

use same env. variables everywhere

same env var for all

separated tests with real mirrors from mock tests

try this workaround again

Updated the remaining access keys...

debug CI error, works fine on local machine

what about s3server?

test_s3server.py doesn't work yet and s3_mock_keys_set is a dirty workaround

linters

lint

secrets are only on the main branch
