+++ test/s3template.yml
@@ -1,8 +1,8 @@
 mirrors:
   s3test:
     - url: __server__
-      user: env:AWS_ACCESS_KEY
-      password: env:AWS_SECRET_KEY
+      user: env:AWS_ACCESS_KEY_ID
+      password: env:AWS_SECRET_ACCESS_KEY
       region: env:AWS_DEFAULT_REGION
 
 targets:
